+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 3.2.0
+
+- Fix auto-update
+- Add scan interval option (default to 240 minutes)
+- Add option to use brand logo as entity picture or not (default to yes)
+
 ## 3.1.1
 
 - Fix the since days calculation